requirements.txt
# Requirements
pip>=9.0.0
tinydb==3.15.2
PyYAML
PyCLI
ansible>=2.7
ansible-runner
click>=4.0
coloredlogs
netaddr
paramiko>=2.7
schema
subprocess32
ipykernel
jinja2>=2.7
names
sqlitedict
marshmallow
marshmallow_dataclass
pytest
numpy
typeguard

# AWS requirements
pycrypto
boto>=2.48
boto3
botocore

# Other depencencies
ipaddress
scandir
secretstorage<=2.3.1
argcomplete

# Documentation
sphinx
sphinx-autoapi
sphinx_rtd_theme
